control de sesión

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -17,8 +17,8 @@ class CreateUsersTable extends Migration
             $table->bigIncrements('id');
             $table->string('name');
             $table->string('email')->unique();
-            $table->integer('dni');
-            $table->integer('phone');
+            $table->bigInteger('dni');
+            $table->bigInteger('phone');
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->string('rol');
